% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omb_eurofins.R
\name{omb_eurofins}
\alias{omb_eurofins}
\title{Calculate the mean decomposition rate of Soil Organic Matter}
\usage{
omb_eurofins(
  B_SOILTYPE_AGR,
  A_SOM_LOI = NA,
  A_C_OF = NA,
  A_N_RT = NA,
  A_PH_CC = NA,
  A_N_PMN = NA,
  duration = 10
)
}
\arguments{
\item{B_SOILTYPE_AGR}{(character) The agricultural type of soil}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}

\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_N_RT}{(numeric) The organic nitrogen content of the soil in mg N / kg}

\item{A_PH_CC}{(numeric) The pH-CaCl2 of the soil}

\item{A_N_PMN}{(numeric) The potentially mineralizable N pool (mg N / kg soil)}

\item{duration}{(numeric) The duration over which an an annual averaged decomposition rate is estimated (years). Default value is 10 years.}
}
\description{
Calculate the mean decomposition rate of Soil Organic Matter
}
\references{
Eurofins Agro (2013). Formules om OS-balans te berekenen voor NLV. Extension made by Hanegraaf & Bussink (2007) Schatting afbraaksnelheid organische stof. Vertaling van empirisch onderzoek naar nieuwe vuistregels - update 2007. NMI report B112.08, 15 pp.
}
