% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obm_vermeulen.R
\name{omb_vermeulen}
\alias{omb_vermeulen}
\title{Calculate the mean decomposition rate of Soil Organic Matter}
\usage{
omb_vermeulen(
  A_SOM_LOI = NA_real_,
  A_C_OF = NA_real_,
  A_N_RT = NA_real_,
  duration = 10
)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}

\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_N_RT}{(numeric) The total nitrogen content of the soil (mg / kg).}

\item{duration}{(numeric) The duration over which an an annual averaged decomposition rate is estimated (years). Default value is 10 years.}
}
\description{
Calculate the mean decomposition rate of Soil Organic Matter
}
\references{
Vermeulen & Hendriks (1996).Bepaling van afbraaksnelheden van organische stof in laagveen: ademhalingsmetingen aan ongestoorde veenmonsters in het laboratorium. DLO-Staring centrum, Wageningen, Rapport 288, 99 pp.
}
