% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omb_zwart.R
\name{omb_zwart}
\alias{omb_zwart}
\title{Calculate the mean decomposition rate of Soil Organic Matter}
\usage{
omb_zwart(A_SOM_LOI = NA, A_C_OF = NA, A_CLAY_MI = 5)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}

\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}
}
\description{
Calculate the mean decomposition rate of Soil Organic Matter
}
\references{
Zwart et al. (2013) De organische stof balans met de te verwachten stikstoflevering per teeltrotatie. Available at: https://edepot.wur.nl/272649.
}
